requirements.txt
psycopg2-binary==2.9.10
beautifulsoup4
lxml
pytesseract
Pillow
pdfminer.six
regex
pyyaml
numpy
PyPDF2==3.0.1
openai